Terrain
Worked on terrain and did something

+++ Pixel.h
@@ -1,8 +1,8 @@
 #pragma once
 struct Pixel 
 {
-	int r;
-	int g;
-	int b;
-	int a;
+	float r;
+	float g;
+	float b;
+	float a;
 };
